requirements.txt
django==3.1.3
django-autoslug==1.9.8
django-debug-toolbar==3.1.1
django-model-utils==4.1.0
django-crispy-forms==1.12.0
factory-boy==3.1.0
openpyxl==3.0.7
pillow==8.0.1
ipython==7.19.0
psycopg2==2.8.6
pytest-cov==2.10.1
pytest-django==4.1.0
pytest==6.1.2
